feat(build): Use code splitting to dedupe main & core entries
