+++ core.js
@@ -1,3 +1,3 @@
 /* eslint import/no-unresolved:0 */
 // this file just makes it easier to require dist/core
-module.exports = require('./dist/rtl-css-js.core.cjs')
+module.exports = require('./dist/cjs/core.js')
