--- src/core/index.js
@@ -1,4 +1,0 @@
-import propertyValueConverters from './property-value-converters'
-
-export * from './utils'
-export {propertyValueConverters}
